Final file add
To anyone reading this, I've been struggling to find the correct locations of these files.  Turns out there were three versions of the homework on my machine, resulting in a ton of confusion.  I'm working on one set of code now.  If the code appears to be copy and pasted, it was, from my other two versions where changes were made.